+++ server/api/controllers/oauth.controller.ts
@@ -1,10 +1,14 @@
+import { IGitHubService } from "../external-services/github.service";
 import { IGoogleService } from "../external-services/google.service";
 import { HttpStatus } from "../lib/constant/http.type";
 import { ApiResponse } from "../lib/helpers/api-response";
 import { MyError } from "../lib/helpers/errors";
 import { Utils } from "../lib/helpers/utils";
 import { CreateFactoryType } from "../lib/types/factory.type";
-import { GoogleValidation } from "../lib/validations/schema.validation";
+import {
+    GitHubValidation,
+    GoogleValidation,
+} from "../lib/validations/schema.validation";
 import { Validator } from "../lib/validations/validator";
 import { zValidator } from "@hono/zod-validator";
 import { getCookie, setCookie } from "hono/cookie";
@@ -21,10 +25,11 @@ export class OauthController implements IOauthController {
     constructor(
         private readonly factory: CreateFactoryType,
         private readonly googleSerivce: IGoogleService,
+        private readonly gitHubSerivce: IGitHubService,
     ) {
         this.cookiesName = {
-            GOOGLE_STATE: "google_oauth_state",
-            GOOGLE_CODE_VERIFIER: "google_oauth_codeVerifier",
+            OAUTH_STATE: "oauth_state",
+            OAUTH_CODE_VERIFIER: "oauth_codeVerifier",
         } as const;
     }
     public setupHandlers() {
@@ -32,18 +37,20 @@ export class OauthController implements IOauthController {
             .createApp()
             .basePath("/oauth")
             .get("/google", ...this.loginGoogleHandler())
-            .get("/google/callback", ...this.loginGoogleCallbackHandler());
+            .get("/google/callback", ...this.loginGoogleCallbackHandler())
+            .get("/github", ...this.loginGitHubHandler())
+            .get("/github/callback", ...this.loginGitHubCallbackHandler());
     }
     private loginGoogleHandler() {
         return this.factory.createHandlers(async (c) => {
             const { state, url, codeVerifier } =
                 await this.googleSerivce.getOauthConsentUrl();
-            setCookie(c, this.cookiesName.GOOGLE_STATE, state, {
+            setCookie(c, this.cookiesName.OAUTH_STATE, state, {
                 httpOnly: true,
                 maxAge: 60 * 10,
                 secure: envClient.NODE_ENV === "production",
             });
-            setCookie(c, this.cookiesName.GOOGLE_CODE_VERIFIER, codeVerifier, {
+            setCookie(c, this.cookiesName.OAUTH_CODE_VERIFIER, codeVerifier, {
                 httpOnly: true,
                 maxAge: 60 * 10,
                 secure: envClient.NODE_ENV === "production",
@@ -68,9 +75,9 @@ export class OauthController implements IOauthController {
             async (c) => {
                 const { code, state } = c.req.valid("query");
                 const cookieCodeVerifier =
-                    getCookie(c, this.cookiesName.GOOGLE_CODE_VERIFIER) ?? null;
+                    getCookie(c, this.cookiesName.OAUTH_CODE_VERIFIER) ?? null;
                 const cookieState =
-                    getCookie(c, this.cookiesName.GOOGLE_STATE) ?? null;
+                    getCookie(c, this.cookiesName.OAUTH_STATE) ?? null;
                 console.log("cookieState", cookieState);
                 if (
                     !cookieCodeVerifier ||
@@ -102,6 +109,98 @@ export class OauthController implements IOauthController {
                 if (!session || !sessionCookie) {
                     throw new MyError.UnauthorizedError(
                         "Something went wrong with your Google Account",
+                    );
+                }
+                setCookie(c, sessionCookie.name, sessionCookie.value, {
+                    ...sessionCookie.attributes,
+                    sameSite: "Strict",
+                });
+                return ApiResponse.WriteJSON({
+                    c,
+                    status: HttpStatus.OK,
+                    msg: "Sign in successfully",
+                    data: undefined,
+                });
+            },
+        );
+    }
+    private loginGitHubHandler() {
+        return this.factory.createHandlers(async (c) => {
+            const { state, url } =
+                await this.gitHubSerivce.getOauthConsentUrl();
+            console.log("state", state);
+            setCookie(c, this.cookiesName.OAUTH_STATE, state, {
+                httpOnly: true,
+                maxAge: 60 * 10,
+                secure: envClient.NODE_ENV === "production",
+            });
+            return ApiResponse.WriteJSON({
+                c,
+                status: 200,
+                data: {
+                    redirectTo: url.toString(),
+                },
+                msg: "Redirect",
+            });
+        });
+    }
+    private loginGitHubCallbackHandler() {
+        const queries = z.object({
+            state: z.string().min(1),
+            code: z.string().min(1),
+        });
+        return this.factory.createHandlers(
+            zValidator("query", queries, Validator.handleParseError),
+            async (c) => {
+                const { code, state } = c.req.valid("query");
+                const cookieState =
+                    getCookie(c, this.cookiesName.OAUTH_STATE) ?? null;
+                if (!cookieState || cookieState !== state) {
+                    console.error("No state");
+                    throw new MyError.BadRequestError();
+                }
+                const { accessToken } =
+                    await this.gitHubSerivce.verifyCode(code);
+                const gitHubResponse = await fetch(envServer.GITHUB_API_URL, {
+                    headers: {
+                        Authorization: `Bearer ${accessToken}`,
+                    },
+                });
+                const gitHubEmailResponse = await fetch(
+                    `${envServer.GITHUB_API_URL}/emails`,
+                    {
+                        headers: {
+                            Authorization: `Bearer ${accessToken}`,
+                        },
+                    },
+                );
+                if (!gitHubResponse.ok || !gitHubEmailResponse) {
+                    throw new MyError.BadRequestError(
+                        "Failed to fetch GitHub Information",
+                    );
+                }
+                const gitHubData = await gitHubResponse.json();
+                const gitHubEmail = await gitHubEmailResponse.json();
+
+                if (!gitHubEmail[0].email) {
+                    throw new MyError.BadRequestError(
+                        "Github's email is not public, please public it and revoke our authorize to continue",
+                    );
+                }
+
+                const gitHubDataS = GitHubValidation.responseSchema.parse({
+                    id: gitHubData.id,
+                    email: gitHubEmail[0].email,
+                    name: gitHubData.name || gitHubData.login,
+                    avatar_url: gitHubData.avatar_url,
+                    verified_email: true,
+                });
+                console.log("CONTRLLER");
+                const { session, sessionCookie } =
+                    await this.gitHubSerivce.authenticateUser(gitHubDataS);
+                if (!session || !sessionCookie) {
+                    throw new MyError.UnauthorizedError(
+                        "Something went wrong with your GitHub Account",
                     );
                 }
                 setCookie(c, sessionCookie.name, sessionCookie.value, {
